backend/requirements.txt
django
python-dotenv
google-generativeai